pyproject.toml
[project]
name = "django-biology"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Denis Shumeev", email = "denis-shumeev@yandex.ru"},
]
dependencies = [
    "django>=5.0",
    "gunicorn>=21.2.0",
    "pytest>=7.4.3",
    "redis>=5.0.1",
    "sentry-sdk>=1.39.1",
    "markdown>=3.5.1",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "djangorestframework>=3.14.0",
    "django-debug-toolbar>=4.2.0",
    "django-cachalot>=2.3.5",
    "django-cleanup>=8.0.0",
    "django-widget-tweaks>=1.5.0",
    "python-slugify>=8.0.1",
    "pillow>=10.1.0",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
